Disallow git commands from within unpopulated submodules
The main motivations of for disallowing git commands within an
unpopulated submodule are :

Whenever we run "git -C status" within an unpopulated submodule, it
falls back to the superproject. This occurs since there is no .git
file in the , hence the setup_git_dir_gently_1 falls back to the
superproject. So superproject's status gets displayed.

When we run the command "git -C sub add ." within a submodule, the
results observed are:

1. In case of populated submodule, it acts like running “git add .“
inside the submodule. This is uncontroversial and runs as expected.

2. In case of unpopulated submodule, the user's intention behind
entering the above command is unclear. He may have intended to add
the submodule to the superproject or to add all files inside the
sub/ directory to the submodule or superproject. Hence we’ll prefer
out in these case.

Eventually, we use a check_prefix_inside_submodule to see check if the
path is inside an unpopulated submodule. If it is , then we report the
user about the unpopulated submodule.

Signed-off-by: Prathamesh Chavan <pc44800@gmail.com>