submodule: port submodule subcommand sync from shell to C
The submodule subcommand 'sync' is ported to submodule--helper, and
submodule--helper is called from git-submodule.sh. The code is split
up to have one function to obtain all the list of submodules. This
function acts as the front-end of git-submodule sync subcommand. It
calls the function for_each_submodule_list, which basically loops
through the list and calls function fn, which in this case is
sync_submodule. This third function is a calling function that takes
care of synchronizing that submodule, and recursively perform the
same when --recursive is flagged.

The first function module_sync first parses the options present in
argv, and then with the help of module_list_compute, generates the
list of submodules present in the current working tree.

The second function for_each_submodule_list traverses through the
list, and calls function fn (which in case of submodule subcommand
sync is sync_submodule) is called for each entry.

The third function sync_submodule takes care of the recursive flag,
by creating a separate child_process structure and prepending
"--super-prefix displaypath" to the args argv_array structure.

Mentored-by: Christian Couder <christian.couder@gmail.com>
Mentored-by: Stefan Beller <sbeller@google.com>
Signed-off-by: Prathamesh Chavan <pc44800@gmail.com>