dir: create function count_slashes
Similar functions exist in apply.c and builtin/show-branch.c for counting
the number of slashes in a string. Also in the later patches, we introduce
a third caller for the same. Hence, we unify it now by cleaning the existing
functions and declaring a common function count_slashes in dir.h and
implementing it in dir.c to remove this code duplication.

Mentored-by: Christian Couder <christian.couder@gmail.com>
Mentored-by: Stefan Beller <sbeller@google.com>
Signed-off-by: Prathamesh Chavan <pc44800@gmail.com>